fixed type defs
